Introduce TracePos type for trace positions (#273)
Many internal structures refer to trace positions; this is an important type of values. In the code, it is sometimes `u32`, sometimes `usize`. While such variables usually have "_pos" suffix, having a dedicated type does improve code with type guarantees.

+++ air/tests/test_module/features/data_merging/data_merge.rs
@@ -300,7 +300,7 @@ fn fold_merge() {
 
         if let ExecutedState::Fold(fold) = state {
             for subtrace_lore in fold.lore.iter() {
-                let value_pos = subtrace_lore.value_pos as usize;
+                let value_pos = subtrace_lore.value_pos;
                 if let ExecutedState::Call(CallResult::Executed(value)) = &data.trace[value_pos] {
                     let value = match value {
                         Value::Scalar(value) => value,
